refactor: fetchApi 함수 apis 폴더로 이동
