+++ src/apis/index.js
@@ -1,5 +1,5 @@
 import { popularUrl, searchUrl } from "../constants/urls";
-import fetchApi from "../utils/fetchApi";
+import fetchApi from "./fetchApi";
 
 export const fetchMovieListWithKeyword = async (pageNumber, keyword) => {
   const url = `${searchUrl}?api_key=${process.env.WEBPACK_API_KEY}&query=${keyword}&language=ko&page=${pageNumber}`;
